requirements.txt
Flask
psycopg2-binary
gunicorn
pandas
numpy
seaborn
matplotlib
statsmodels
sklearn
radon
pyyaml